LayoutTests/inspector/console/console-repeat-count-expected.txt

CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
CONSOLE MESSAGE: line 10: Same message
Tests that repeat count is properly updated.

5console-repeat-count.html:10Same message
"Different message"
"Different message"
5console-repeat-count.html:10Same message
